完善了Calendar的请求操作和相互跳转

+++ app/src/main/java/com/lire/utils/GsonExtends.kt
@@ -24,7 +24,7 @@ fun JsonElement.getAsNullableInt() : Int? {
     return asInt
 }
 
-fun JsonElement.getAsNullableDouble() : Int? {
+fun JsonElement.getAsNullableDouble() : Double? {
     if (isJsonNull) return null
-    return asInt
+    return asDouble
 }
